Fixes to allow precompiling of assets
sass-rails needed an update (which wouldn't work with a simple `bundle
update sass`, not sure why) and wouldn't you know it, the asset pipeline
is kinda stupid. Or sass is. If A.sass defines a mixin called 'a', is imported into
B.sass, and then C.sass is also imported into B.sass, C.sass cannot use
mixin 'a'. This is likely fixed in Rails 3.2 or 4.0, but for now we
can't clean up the stylesheets like we'd like to.