Test (and clean up) every route. Basically.
The routes were a little sloppy here and there. This ensures that all
controllers can be reached at the desired URLs.

Note: This does not necessarily ensure that "publications_path" goes to
"/", though, just that "/" goes to PublicationsController#index.
However, cucumber features and request specs managed to highlight
some problems of the latter type.