Remove dependency on ADMIN_EMAIL and EDITOR_EMAIL env vars
This required changing the logic for acquiring the current editor.
Previously, the current editor was a global setting, and finding it was
messy. Now, there's a nice `Publication#editor` method to return it.

However, adding and using this caused ~95% of the specs to break. Fixing
them required much effort and investigation.

I've also removed many references to "Problem Child" in the source of
the app, though much still remains. Removing it should be somewhat
trivial, in comparison to this change.