Only load magazines for the current publication
Cucumber is huge and bulky so I decided to test this with RSpec request
specs. Unfortunately, it seems that sessions aren't working in my
request specs. I spent way too long trying to track down why, then
decided to just stub the sign in. If I ever decide to move the testing
of the sign in page to the request specs and out of Cucumber, this will
need to be dealt with at that time. But for all other sign in needs,
this will probably work.