Scope magazine timeframe logic to current publication
A few tests are failing because of these changes. I started attempting
to fix them all now, and as a result, this has also been finished:

* Ensure new magazines are created with a publication

However, some specs & features are still failing. But this commit was
starting to get away from me, so I'm pushing now, then cleaning up the
remaining failures.