Move some controller logic into "Homepage" object
This was mostly an intellectual exercise to attempt to follow Sandi
Metz's "one instantiation per controller action" rule. I've obviously
failed to follow that rule, even still. My reason:

`@publication` is assumed to be set in the layout, and short of
refactoring the whole site, it needs to be set.

As far as moving the other logic into a `Homepage` object, it seems to
have accomplished a few nice things:

* the logic for what the homepage needs is contained in one
  outside-the-Rails-stack, easy-to-test place (I'm not currently taking
  advantage of this)
* stubbing this new object to keep existing tests passing was easy
* the view cleanly declares what data certain sections are interested
  in, which is interesting. I'm not sure I care much, since html
  sections & ids already accomplished that.