Test drive the new home page with real data
And adjust accordingly. Lessons learned:

* The home page should only feature submissions from fully-published
  magazines; that is, magazines that have had a notification sent
* Truncating data that contains html tags is bad news
* Including all of the contents on the page & limiting the height with
  CSS is nbd
* Photographs looks really dumb if they're faded out (will need to deal
  with this more intelligently later, instead of just not including
  them)

The HookSubmission thus went away, which was nice.

The definition of @hook in lib/homepage is a little bit goofy, mainly
because during testing the publication may not have any published
magazines, and it needs to account for that. Someday, when people first
sign up for Publishist, this will also be the case, so the home page
really ought to account for a blank @hook.